+++ lib/widgets/habits/habits_filter.dart
@@ -7,6 +7,7 @@ import 'package:lotti/database/database.dart';
 import 'package:lotti/get_it.dart';
 import 'package:lotti/utils/color.dart';
 import 'package:pie_chart/pie_chart.dart';
+import 'package:tinycolor2/tinycolor2.dart';
 
 class HabitsFilter extends StatelessWidget {
   const HabitsFilter({super.key});
@@ -26,6 +27,7 @@ class HabitsFilter extends StatelessWidget {
         return BlocBuilder<HabitsCubit, HabitsState>(
           builder: (context, HabitsState state) {
             final dataMap = <String, double>{};
+            final cubit = context.read<HabitsCubit>();
 
             for (final habit in state.openNow) {
               final categoryId = habit.categoryId ?? 'undefined';
@@ -40,26 +42,76 @@ class HabitsFilter extends StatelessWidget {
                   : Colors.grey;
             }).toList();
 
-            if (dataMap.isEmpty) {
-              return const SizedBox.shrink();
-            }
-
-            return Padding(
+            return IconButton(
               padding: const EdgeInsets.all(5),
-              child: PieChart(
-                dataMap: dataMap,
-                animationDuration: const Duration(milliseconds: 800),
-                chartRadius: 25,
-                colorList: colorList,
-                initialAngleInDegree: 0,
-                chartType: ChartType.ring,
-                ringStrokeWidth: 10,
-                legendOptions: const LegendOptions(showLegends: false),
-                chartValuesOptions: const ChartValuesOptions(
-                  showChartValueBackground: false,
-                  showChartValues: false,
-                ),
-              ),
+              icon: dataMap.isEmpty
+                  ? const Icon(Icons.filter_alt)
+                  : PieChart(
+                      dataMap: dataMap,
+                      animationDuration: const Duration(milliseconds: 800),
+                      chartRadius: 25,
+                      colorList: colorList,
+                      initialAngleInDegree: 0,
+                      chartType: ChartType.ring,
+                      ringStrokeWidth: 10,
+                      legendOptions: const LegendOptions(showLegends: false),
+                      chartValuesOptions: const ChartValuesOptions(
+                        showChartValueBackground: false,
+                        showChartValues: false,
+                      ),
+                    ),
+              onPressed: () {
+                showModalBottomSheet<void>(
+                  context: context,
+                  useRootNavigator: true,
+                  builder: (BuildContext context) {
+                    return BlocProvider.value(
+                      value: cubit,
+                      child: BlocBuilder<HabitsCubit, HabitsState>(
+                        builder: (context, HabitsState state) {
+                          return Container(
+                            padding: const EdgeInsets.symmetric(
+                              vertical: 20,
+                              horizontal: 10,
+                            ),
+                            child: Wrap(
+                              spacing: 5,
+                              runSpacing: 5,
+                              children: [
+                                ...categories.map((category) {
+                                  final color = colorFromCssHex(category.color);
+
+                                  return Opacity(
+                                    opacity: state.selectedCategoryIds
+                                            .contains(category.id)
+                                        ? 1
+                                        : 0.4,
+                                    child: ActionChip(
+                                      onPressed: () =>
+                                          cubit.toggleSelectedCategoryIds(
+                                        category.id,
+                                      ),
+                                      label: Text(
+                                        category.name,
+                                        style: TextStyle(
+                                          color: color.isLight
+                                              ? Colors.black
+                                              : Colors.white,
+                                        ),
+                                      ),
+                                      backgroundColor: color,
+                                    ),
+                                  );
+                                })
+                              ],
+                            ),
+                          );
+                        },
+                      ),
+                    );
+                  },
+                );
+              },
             );
           },
         );